backend/requirements.txt
aioboto3==9.5.0
aiobotocore==2.2.0
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
boto3==1.21.21
botocore==1.24.21
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
css-inline==0.8.1
cssselect==1.1.0
cssutils==2.3.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
demjson==2.2.4
dnspython==2.2.0
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
fastjsonschema==2.15.3
Flask==2.0.2
Flask-Admin==1.6.0
Flask-CKEditor==0.4.6
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
frozenlist==1.3.0
h11==0.12.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
json5==0.9.6
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
minify_html==0.8.0
mistune==0.8.4
mongoengine==0.23.1
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.3.0
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.0
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyactiveresource==2.2.2
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyJWT==2.3.0
Pyliner==1.0
pymongo==3.12.3
pynliner==0.8.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.5
pytest-asyncio==0.18.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.27.1
s3transfer==0.5.1
selenium==4.1.0
Send2Trash==1.8.0
ShopifyAPI==9.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
stack-data==0.1.4
starlette==0.17.1
terminado==0.13.1
testpath==0.5.0
toml==0.10.2
tomli==2.0.0
tornado==6.1
toronado==0.1.0
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webdriver-manager==3.5.2
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.14.1
wsproto==1.0.0
WTForms==3.0.1
xlrd==2.0.1
yarl==1.7.2
pygifsicle